requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
psycopg2-binary
redis
kafka-python
pyyaml
gitpython
openai
langchain
weaviate-client
langchain-openai
google-search-results
serpapi